Fix double rendering of item tooltips with chat triggers
The root cause here is
https://github.com/ChatTriggers/ChatTriggers/blob/master/src/main/kotlin/com/chattriggers/ctjs/launch/plugin/guiContainer.kt
not separating their code into its own method, so the original injection
point would find their popMatrix attribute as well as the vanilla one. I
changed this to now use a different injection point which is not being
used by chattriggers. The real fix would be to just move the
chattriggers code into a hook method.